KAFKA-13075: Consolidate RocksDBStore and RocksDBKeyValueStoreTest
